Move StoreAttachment into org class
