+++ models/orgs_test.go
@@ -1,13 +1,16 @@
 package models
 
 import (
+	"io/ioutil"
 	"testing"
 	"time"
 
 	"github.com/nyaruka/goflow/envs"
+	"github.com/nyaruka/goflow/utils"
 	"github.com/nyaruka/mailroom/testsuite"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestOrgs(t *testing.T) {
@@ -52,4 +55,23 @@ func TestOrgs(t *testing.T) {
 
 	_, err = loadOrg(ctx, tx, 99)
 	assert.Error(t, err)
+}
+
+func TestStoreAttachment(t *testing.T) {
+	ctx := testsuite.CTX()
+	db := testsuite.DB()
+
+	store := testsuite.Storage()
+	defer testsuite.ResetStorage()
+
+	image, err := ioutil.ReadFile("testdata/test.jpg")
+	require.NoError(t, err)
+
+	org, err := loadOrg(ctx, db, Org1)
+	assert.NoError(t, err)
+
+	attachment, err := org.StoreAttachment(store, "media", "668383ba-387c-49bc-b164-1213ac0ea7aa.jpg", image)
+	require.NoError(t, err)
+
+	assert.Equal(t, utils.Attachment("image/jpeg:_test_storage/media/1/6683/83ba/668383ba-387c-49bc-b164-1213ac0ea7aa.jpg"), attachment)
 }
